+++ src/authenticateUser/index.d.ts
@@ -1,4 +1,5 @@
 import type { session_user_ctx_T } from '../_types'
 export declare function authenticateUser(
-	Username:string, Password:string
+	Username:string,
+	Password:string
 ):Promise<session_user_ctx_T>