+++ src/changePassword/index.d.ts
@@ -1,4 +1,6 @@
 import type { CognitoUser } from 'amazon-cognito-identity-js'
 export declare function changePassword(
-	user:CognitoUser, oldPassword:string, newPassword:string
+	user:CognitoUser,
+	oldPassword:string,
+	newPassword:string
 ):Promise<'SUCCESS'>