+++ src/changePassword/index.js
@@ -1,11 +1,15 @@
 import { promisify } from 'util'
 /**
- * @param user{module:amazon-cognito-identity-js.CognitoUser}
- * @param oldPassword{string}
- * @param newPassword{string}
+ * @param {module:amazon-cognito-identity-js.CognitoUser}user
+ * @param {string}oldPassword
+ * @param {string}newPassword
  * @returns {Promise<"SUCCESS"|undefined>}
  */
-export async function changePassword(user, oldPassword, newPassword) {
+export async function changePassword(
+	user,
+	oldPassword,
+	newPassword
+) {
 	const changePassword_ = promisify(user.changePassword.bind(user))
 	return await changePassword_(oldPassword, newPassword)
 }