+++ src/confirmRegistration/index.d.ts
@@ -1,3 +1,4 @@
 export declare function confirmRegistration(
-	Username:string, code:string
+	Username:string,
+	code:string
 ):Promise<any>