+++ src/deleteAttributes/index.d.ts
@@ -1,4 +1,5 @@
 import type { CognitoUser } from 'amazon-cognito-identity-js'
 export declare function deleteAttributes(
-	user:CognitoUser, attribute_a:string[]
+	user:CognitoUser,
+	attribute_a:string[]
 ):Promise<string>