+++ src/deleteAttributes/index.js
@@ -1,10 +1,13 @@
 import { promisify } from 'util'
 /**
- * @param user{module:amazon-cognito-identity-js.CognitoUser}
- * @param attribute_a{string[]}
+ * @param {module:amazon-cognito-identity-js.CognitoUser}user
+ * @param {string[]}attribute_a
  * @returns {Promise<string>}
  */
-export async function deleteAttributes(user, attribute_a) {
+export async function deleteAttributes(
+	user,
+	attribute_a
+) {
 	const deleteAttributes_ = promisify(user.deleteAttributes.bind(user))
 	return await deleteAttributes_(attribute_a)
 }