+++ src/deleteUser/index.js
@@ -1,6 +1,6 @@
 import { promisify } from 'util'
 /**
- * @param user{module:amazon-cognito-identity-js.CognitoUser}
+ * @param {module:amazon-cognito-identity-js.CognitoUser}user
  * @returns {Promise<string>}
  */
 export async function deleteUser(user) {