+++ src/getAttributeVerificationCode/index.d.ts
@@ -1,5 +1,6 @@
 import type { CognitoUser } from 'amazon-cognito-identity-js'
 import type { inputVerificationCode_fn_T } from '../_types'
 export declare function getAttributeVerificationCode(
-	user:CognitoUser, inputVerificationCode:inputVerificationCode_fn_T
+	user:CognitoUser,
+	inputVerificationCode:inputVerificationCode_fn_T
 ):Promise<null>