+++ src/getUserAttributes/index.js
@@ -1,10 +1,13 @@
 /**
- * @param user{module:amazon-cognito-identity-js.CognitoUser}
+ * @param {module:amazon-cognito-identity-js.CognitoUser}user
  * @returns {Promise<module:amazon-cognito-identity-js.CognitoUserAttribute[]>}
  */
 export async function index(user) {
 	return new Promise((resolve, reject)=>{
-		user.getUserAttributes((err, result)=>{
+		user.getUserAttributes((
+			err,
+			result
+		)=>{
 			if (err) {
 				reject(err)
 				return