+++ src/globalSignOut/index.d.ts
@@ -1,2 +1,4 @@
 import type { CognitoUser } from 'amazon-cognito-identity-js'
-export declare function globalSignOut(user:CognitoUser):Promise<string>
+export declare function globalSignOut(
+	user:CognitoUser
+):Promise<string>