+++ src/signUp/index.js
@@ -2,10 +2,10 @@ import { promisify } from 'util'
 import { Pool } from '../Pool/index.js'
 const signUp_ = promisify(Pool.signUp.bind(Pool))
 /**
- * @param username{string}
- * @param password{string}
- * @param attributes_a{module:amazon-cognito-identity-js.CognitoUserAttribute[]}
- * @param validation_data{module:amazon-cognito-identity-js.CognitoUserAttribute[]}
+ * @param {string}username
+ * @param {string}password
+ * @param {module:amazon-cognito-identity-js.CognitoUserAttribute[]}attributes_a
+ * @param {module:amazon-cognito-identity-js.CognitoUserAttribute[]}validation_data
  * @returns {Promise<ISignUpResult>}
  */
 export async function signUp(